Inline board methods
